Create setup.py
get_user_id Function:

Defines the path to the config.json configuration file.
If the file doesn't exist, returns the string "anonymous_user".
If the file exists, attempts to read it and obtain the user_id value. If there is an error during the read, it also returns "anonymous_user".